Fix way more clangd warnings (#105)
* More clangd fixes

* Remove clang format around FORCE macros

* Uniform TODO comments

* Replace ifdef/ifndef with if

* Convert multi-character constants

* Fix more warnings

+++ src/egg/core/eggDisplay.cpp
@@ -1,4 +1,4 @@
-// TODO: REMOVE AFTER REFACTOR
+// TODO(kiwi) REMOVE AFTER REFACTOR
 #pragma ipa file
 
 #include <egg/core.h>
@@ -82,7 +82,9 @@ void Display::calcFrequency() {
     s32 nowFrameTick = OSGetTick();
     mFrameCostTick = nowFrameTick - mPrevFrameTick;
 
-    mFrequency = 1000000.0f / OS_TICKS_TO_USEC(mFrameCostTick);
+    mFrequency =
+        1000000.0f / static_cast<u32>(OS_TICKS_TO_USEC(mFrameCostTick));
+
     mPrevFrameTick = nowFrameTick;
 }
